moved where themeawaresocialicon is
