+++ clocktower/.env.local.example
@@ -4,4 +4,3 @@ NEXT_PUBLIC_SUPABASE_ANON_KEY=your-supabase-anon-key # Required
 NEXT_PUBLIC_PROVIDERS='' # This is used for authentication providers, it is totally optional and right now unsupported
 SUPABASE_SERVICE_KEY=yoursupersecretservicekey # Required Service key used for service operations
 OPENAI_API_KEY=your-openai-api-key #This is used for moderation of usernames, it is totally optional
-LOCAL_ONLY_TESTS=true # This is used for the testing suite, if set to true it will mock api calls, if set to false it will use the actual api
